ngzblog/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
baidu-aip==2.2.8.0
beautifulsoup4==4.6.3
billiard==3.5.0.4
bs4==0.0.1
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
diff-match-patch==20121119
Django==1.11.14
django-appconf==1.0.2
django-bootstrap3==11.0.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
Pillow==5.3.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.4
requests==2.19.1
six==1.10.0
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
